backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# LangGraph & LLM
langchain==0.1.0
langgraph==0.0.20
langchain-community==0.0.13
ollama==0.1.7

# PyTorch
torch==2.1.1
torchvision==0.16.1
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2

# Financial Data
yfinance==0.2.32
requests==2.31.0

# Redis
redis==5.0.1
hiredis==2.2.3

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
python-dateutil==2.8.2

# Logging & Monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.39.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

